refactor: move go project to repo root
