+++ .gitignore
@@ -1,4 +1,4 @@
-TrendRader_go/bin/
-TrendRader_go/output/
+bin/
+output/
 __pycache__/
 *.pyc
