+++ cmd/trendradar/main.go
@@ -8,11 +8,11 @@ import (
 	"os"
 	"time"
 
-	"github.com/sansan0/TrendRadar/go/pkg/analysis"
-	"github.com/sansan0/TrendRadar/go/pkg/app"
-	"github.com/sansan0/TrendRadar/go/pkg/config"
-	"github.com/sansan0/TrendRadar/go/pkg/crawler"
-	"github.com/sansan0/TrendRadar/go/pkg/storage"
+	"github.com/sansan0/TrendRadar/pkg/analysis"
+	"github.com/sansan0/TrendRadar/pkg/app"
+	"github.com/sansan0/TrendRadar/pkg/config"
+	"github.com/sansan0/TrendRadar/pkg/crawler"
+	"github.com/sansan0/TrendRadar/pkg/storage"
 )
 
 func main() {
