+++ docker/docker-compose-go.yml
@@ -2,7 +2,7 @@ services:
   trendradar-go:
     build:
       context: ..
-      dockerfile: TrendRader_go/Dockerfile
+      dockerfile: Dockerfile
     image: trendradar-go:latest
     container_name: trendradar-go
     environment:
@@ -10,14 +10,14 @@ services:
       - CONFIG_PATH=/app/config/config.yaml
       - FREQUENCY_WORDS_PATH=/app/config/frequency_words.txt
     volumes:
-      - ../TrendRader_go/config:/app/config:ro
-      - ../TrendRader_go/output:/app/output
+      - ../config:/app/config:ro
+      - ../output:/app/output
     restart: unless-stopped
 
   mcp-go:
     build:
       context: ..
-      dockerfile: TrendRader_go/Dockerfile
+      dockerfile: Dockerfile
     image: trendradar-go:latest
     container_name: trendradar-go-mcp
     environment:
@@ -25,6 +25,6 @@ services:
       - CONFIG_PATH=/app/config/config.yaml
       - FREQUENCY_WORDS_PATH=/app/config/frequency_words.txt
     volumes:
-      - ../TrendRader_go/config:/app/config:ro
-      - ../TrendRader_go/output:/app/output
+      - ../config:/app/config:ro
+      - ../output:/app/output
     restart: unless-stopped
