+++ pkg/mcptools/config_tool.go
@@ -1,6 +1,6 @@
 package mcptools
 
-import "github.com/sansan0/TrendRadar/go/pkg/app"
+import "github.com/sansan0/TrendRadar/pkg/app"
 
 func CurrentConfig(env *app.Environment) interface{} {
 	return env.Config
