+++ pkg/mcptools/latest_news.go
@@ -3,7 +3,7 @@ package mcptools
 import (
 	"time"
 
-	"github.com/sansan0/TrendRadar/go/pkg/app"
+	"github.com/sansan0/TrendRadar/pkg/app"
 )
 
 type LatestNewsArgs struct {
