+++ pkg/mcptools/news_by_date.go
@@ -5,7 +5,7 @@ import (
 	"strings"
 	"time"
 
-	"github.com/sansan0/TrendRadar/go/pkg/app"
+	"github.com/sansan0/TrendRadar/pkg/app"
 )
 
 type NewsByDateArgs struct {
