+++ pkg/mcptools/search.go
@@ -6,8 +6,8 @@ import (
 	"strings"
 	"time"
 
-	"github.com/sansan0/TrendRadar/go/pkg/analysis"
-	"github.com/sansan0/TrendRadar/go/pkg/app"
+	"github.com/sansan0/TrendRadar/pkg/analysis"
+	"github.com/sansan0/TrendRadar/pkg/app"
 )
 
 type SearchArgs struct {
